app/requirements.txt
Flask-Migrate-4.0.7
Flask-restx
Flask-sqlalchemy
flask-bcrypt-1.0.1
flask-cors